Put resize code for the ScorePane into its own class
