+++ src/main/java/ui/controller/ScorePaneController.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright (c) 2017
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package ui.controller;
+
+import javafx.beans.value.ObservableValue;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
+import javafx.scene.control.Separator;
+import javafx.scene.layout.Pane;
+import javafx.scene.layout.VBox;
+
+public class ScorePaneController {
+
+	private Pane scorePane;
+
+	public ScorePaneController(Pane pane) {
+		scorePane = pane;
+		scorePane.widthProperty().addListener(this::resizeScore);
+
+		setVisible(false);
+	}
+
+	public void setVisible(boolean visible) {
+		scorePane.setVisible(visible);
+	}
+
+	private void resizeScore(ObservableValue<? extends Number> observableValue, Number number, Number t1) {
+		VBox blackScoreBox = null;
+		Separator separator = null;
+
+		for (Node node : scorePane.getChildren()) {
+			if ( node instanceof Separator )
+				separator = (Separator) node;
+		}
+
+		if ( separator == null )
+			return;
+
+		for (Node node : scorePane.getChildren()) {
+			if ( node instanceof VBox && node.getLayoutX() < separator.getLayoutX() )
+				blackScoreBox = (VBox) node;
+		}
+
+		if ( blackScoreBox == null )
+			return;
+
+		double width = (scorePane.getWidth() - separator.getWidth()) / 2;
+
+		blackScoreBox.setMinWidth(width);
+	}
+}
